refactor: extract multi-channel adapters into tau-startup
